drivers: serial: stm32 uart align buffers of CM7 with DCACHE
Add support for systems that have a DCACHE. When the buffers
are in cacheable memory the driver will split up the buffers
so the pieces align with the cache lines. The parts that
do not align are first copied to a intermediate buffer that
is correctly aligned. When the user uses memory that is
non-cacheable the option can but turned off for better
performance. Like the SPI driver does.

Signed-off-by: Francois Ramu <francois.ramu@st.com>