drivers: serial: stm32 uart control cache coherency
When UART is using cacheable memory regions for the DMA
source/or destination buffers, it must trigger a cache clean
before starting a DMA operation to ensure that all the data
are committed to the subsystem memory.
After the DMA transfer complete, when reading the data
from the peripheral, it must perform a cache invalidate
before reading the DMA updated memory region.

Signed-off-by: Francois Ramu <francois.ramu@st.com>