drivers: adc: stm32 build the LL_ADC_CLOCK from the DTS property
Get the adc-prescaler property in the DTS to be the ADC clock
asynchronous or synchronous and combine with the divider
to have the exact clock scheme
LL_ADC_CLOCK_ASYNC_DIVx or LL_ADC_CLOCK_SYNC_PCLK_DIVx.
This depends on the stm32 serie (stm32F1 has no prescaler capability).

Signed-off-by: Francois Ramu <francois.ramu@st.com>