boards: arm: vref is 3300 on nucleo_f334 board
The vref is a property of the board, not soc.
On the nucleo_f334R8,  Vref value is 3.3V (see schematics).
It is a common reference to all the ADC.
The DTS is defining and enabling the die-temp sensor
and the ADC1 for temperature measurement.
The ADC1 of the stm32f334 has temp sensor and vref internal channels.

Signed-off-by: Francois Ramu <francois.ramu@st.com>