drivers: spi: stm32 spi with DMA transfer controls the CS pin
Set the CS gpio pin just before the DMA transfer starts.
Then the delay from CS going low and MOSI pin is around 8µs.
With the data->status_flags, the CS GPIO is not set on each loop
but only at the beginning of the DMA activity (stored on one byte).

Signed-off-by: Francois Ramu <francois.ramu@st.com>