drivers: flash: stm32 qspi driver setting NOR flash in MemoryMapped mode
Configures the external NOR Flash in MemoryMapped Mode, at the end of the
NOR flash-controller initialization. Then reading/writing are performed
in MemoryMapped mode with memcopy (and no more with command mode)
In this mode: erasing is not supported anymore.
The flash size and address are given by the DTS <reg> property.
drivers: flash: stm32 qspi memmap config with register access

Signed-off-by: Francois Ramu <francois.ramu@st.com>