boards: arm: stm32u585 disco kit move partition to ext flash
Add the octoSPI 2 which is connection to a octo NOR flash
mx25lm51245 on the stm32u585 disco kit.
Define the partition on the 64MBytes space of the external flash
in the specific dts file (not the _ns one).
Use the STM32Cube programmer to flash with the corresponding external
loader for XiP mode. Do not mass-erase the internal flash when flashing.
Removing usb_device will avoid test feature for that board.

Signed-off-by: Francois Ramu <francois.ramu@st.com>