boards: arm: stm32h7b3 disco kit move partition to ext flash
Add the OSPI 1 which is an octoSPI connection to a octo NOR flash
mx25lm51245 on the stm32u585 disco kit.
Define the partition on the 64MBytes space of the external flash.
Use the STM32Cube programmer to flash with the corresponding external
loader for XiP mode. Do not mass-erase the internal flash when flashing.
Removing usb_device will avoid test feature for that board.

Signed-off-by: Francois Ramu <francois.ramu@st.com>