boards: arm: stm32h7x disco kit move partition to ext flash
Add the quad SPI which is connection to a dual quad NOR flash
mt25ql512ab on the stm32h747 or stm32h750 disco kit.
Put the boot partition in the internal flash.
Define the partition of the external flash with slot0 & 1.
Use the STM32Cube programmer to flash with the corresponding external
loader for XiP mode.
Removing usb_device will avoid test feature for that board.

Signed-off-by: Francois Ramu <francois.ramu@st.com>